+++ src/RPCServer.php
@@ -27,14 +27,13 @@ use Rubix\Server\Http\Controllers\RPCController;
 use Rubix\Server\Http\Middleware\Middleware;
 use Rubix\Server\Serializers\Json;
 use Rubix\Server\Serializers\Serializer;
+use Rubix\Server\Traits\LoggerAware;
 use React\Http\Server as HTTPServer;
 use React\Socket\Server as Socket;
 use React\Socket\SecureServer as SecureSocket;
 use React\EventLoop\Factory as Loop;
 use Psr\Http\Message\ServerRequestInterface as Request;
 use Psr\Http\Message\ResponseInterface as Response;
-use Psr\Log\LoggerAwareInterface as LoggerAware;
-use Psr\Log\LoggerInterface as Logger;
 use InvalidArgumentException;
 
 /**
@@ -47,8 +46,10 @@ use InvalidArgumentException;
  * @package     Rubix/Server
  * @author      Andrew DalPino
  */
-class RPCServer implements Server, LoggerAware
+class RPCServer implements Server, Verbose
 {
+    use LoggerAware;
+
     /**
      * The host address to bind the server to.
      *
@@ -91,13 +92,6 @@ class RPCServer implements Server, LoggerAware
      * @var \Rubix\Server\Http\Controllers\Controller
      */
     protected $controller;
-
-    /**
-     * The logger instance.
-     *
-     * @var Logger|null
-     */
-    protected $logger;
 
     /**
      * The timestamp from when the server went up.
@@ -155,16 +149,6 @@ class RPCServer implements Server, LoggerAware
         $this->cert = $cert;
         $this->middleware = array_values($middleware);
         $this->serializer = $serializer ?? new Json();
-    }
-
-    /**
-     * Sets a psr-3 logger.
-     *
-     * @param Logger|null $logger
-     */
-    public function setLogger(?Logger $logger = null)
-    {
-        $this->logger = $logger;
     }
 
     /**
@@ -244,7 +228,6 @@ class RPCServer implements Server, LoggerAware
     {
         $commands = [
             QueryModel::class => new QueryModelHandler($estimator),
-            ServerStatus::class => new ServerStatusHandler($this),
             Predict::class => new PredictHandler($estimator),
         ];
 
@@ -260,6 +243,10 @@ class RPCServer implements Server, LoggerAware
         if ($estimator instanceof Ranking) {
             $commands[Rank::class] = new RankHandler($estimator);
             $commands[RankSample::class] = new RankSampleHandler($estimator);
+        }
+
+        if ($this instanceof Verbose) {
+            $commands[ServerStatus::class] = new ServerStatusHandler($this);
         }
 
         return new CommandBus($commands);
@@ -287,6 +274,8 @@ class RPCServer implements Server, LoggerAware
             
             $this->logger->info("$status $method $uri from $ip");
         }
+
+        $this->requests++;
 
         return $response;
     }